Исключения и проверка на дубли

+++ Учет цистерн/Forms/Бригады/BrigadeForm.cs
@@ -22,35 +22,49 @@ namespace Учет_цистерн
 
         private void BtnBrigadeUpdate_Click(object sender, EventArgs e)
         {
-            string GetActiveBrigade = "select Active from d__Brigade where ID = " + SelectItemRow;
-            DataTable dataTable = new DataTable();
-            dataTable = DbConnection.DBConnect(GetActiveBrigade);
-            Active = Convert.ToInt32(dataTable.Rows[0][0]);
+            try
+            {
+                string GetActiveBrigade = "select Active from d__Brigade where ID = " + SelectItemRow;
+                DataTable dataTable = new DataTable();
+                dataTable = DbConnection.DBConnect(GetActiveBrigade);
+                Active = Convert.ToInt32(dataTable.Rows[0][0]);
 
-            BrigadeUpdateForm brigadeUpdateForm = new BrigadeUpdateForm
-            {
-                SelectID = SelectItemRow
-            };
-            brigadeUpdateForm.textBox1.Text = dataGVBrigade.CurrentRow.Cells[1].Value.ToString();
-            brigadeUpdateForm.textBox2.Text = dataGVBrigade.CurrentRow.Cells[2].Value.ToString();
-            brigadeUpdateForm.textBox3.Text = dataGVBrigade.CurrentRow.Cells[3].Value.ToString();
-            if (Active == 1)
-            {
-                brigadeUpdateForm.checkBox1.Checked = true;
+                BrigadeUpdateForm brigadeUpdateForm = new BrigadeUpdateForm
+                {
+                    SelectID = SelectItemRow
+                };
+                brigadeUpdateForm.textBox1.Text = dataGVBrigade.CurrentRow.Cells[1].Value.ToString();
+                brigadeUpdateForm.textBox2.Text = dataGVBrigade.CurrentRow.Cells[2].Value.ToString();
+                brigadeUpdateForm.textBox3.Text = dataGVBrigade.CurrentRow.Cells[3].Value.ToString();
+                if (Active == 1)
+                {
+                    brigadeUpdateForm.checkBox1.Checked = true;
+                }
+                else
+                {
+                    brigadeUpdateForm.checkBox1.Checked = false;
+                }
+                brigadeUpdateForm.Show();
             }
-            else
+            catch (Exception ex)
             {
-                brigadeUpdateForm.checkBox1.Checked = false;
+                MessageBox.Show("Для редактирования записи, необходимо указать строку! " + ex.Message);
             }
-            brigadeUpdateForm.Show();
         }
 
         private void BtnBrigadeDelete_Click(object sender, EventArgs e)
         {
-            string Delete = "delete from d__Brigade where ID = " + SelectItemRow;
-            DataTable dataTable = new DataTable();
-            dataTable = DbConnection.DBConnect(Delete);
-            MessageBox.Show("Запись удалена!");
+            string message = "Вы действительно хотите удалить эту запись?";
+            string title = "Удаление";
+            MessageBoxButtons buttons = MessageBoxButtons.OKCancel;
+            DialogResult result = MessageBox.Show(message, title, buttons);
+            if (result == DialogResult.OK)
+            {
+                string Delete = "delete from d__Brigade where ID = " + SelectItemRow;
+                DataTable dataTable = new DataTable();
+                dataTable = DbConnection.DBConnect(Delete);
+                MessageBox.Show("Запись удалена!");
+            }
         }
 
         private void BtnBrigadeReffresh_Click(object sender, EventArgs e)
